pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "alpaca-py>=0.43.2",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "pytz>=2026.1.post1",
    "schedule>=1.2.2",
    "scikit-learn>=1.8.0",
]
